feat(library-solidity): template Solidity  files for ease of dev + update generated code
