+++ library-solidity/codegen/src/paths.ts
@@ -67,3 +67,14 @@ export function isDirectoryEmpty(directoryPath: string): boolean {
   const contents = readdirSync(directoryPath);
   return contents.length === 0;
 }
+
+export function resolveTemplatePath(templateFilename: string): string {
+  // __dirname should be /path/to/fhevm/library-solidity/codegen/dist/_esm
+  // Throw an error in case dist naming is changed or paths.ts is moved somewhere elst
+  // We need to perform the resolve manually
+  if (!__dirname.endsWith('library-solidity/codegen/dist/_esm')) {
+    throw new Error(`Unexpected __dirname value ${__dirname}. Should end with "library-solidity/codegen/dist/_esm"`);
+  }
+
+  return path.normalize(path.join(__dirname, '../../src/templates', templateFilename));
+}
