+++ library-solidity/codegen/src/templates/FheType.sol-template
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: BSD-3-Clause-Clear
+pragma solidity ^0.8.24;
+
+enum FheType {
+    //$$ -----------------------------------------------------------------------
+    //$$ -                                                                     
+    //$$ -          🚚 FheType Enum Values Template Placeholder
+    //$$ -                                                                     
+    //$$ -----------------------------------------------------------------------
+    $${FheTypeEnum}$$
+    //$$ -----------------------------------------------------------------------
+}
